### 2018-03-16 #631 @whedon
[PRE REVIEW]: MLxtend: Providing machine learning and data science utilities and extensions to Python's scientific computing stack
**Submitting author:** @rasbt (<a href="http://orcid.org/0000-0001-6989-4493">Sebastian Raschka</a>)
**Repository:** <a href="https://github.com/rasbt/mlxtend" target ="_blank">https://github.com/rasbt/mlxtend</a>
**Version:** v0.11.0
