### 2018-03-16 b812b37
Update paper.md
