+++ paper.md
@@ -1,6 +1,3 @@
-## MLxtend: Providing machine learning and data science utilities and extensions to Python's scientific computing stack
-
-
 ---
 title: 'MLxtend: Providing machine learning and data science utilities and extensions to Python's scientific computing stack'
 tags:
