### 2018-03-16 #631 @rasbt
@arfon It seems to be fixed now. It was due to a stupid mistake: I enclosed the title in single quotes and had an apostrophe (Python's) in the title.
