### 2018-03-18 #631 @rasbt
Thanks for volunteering as the editor, @arokem !

> Any particular suggestions for reviewers?

@mlgill , @rhiever , @chrisalbon come to mind, who are data science and python experts as well as big open source contributors (and may have not used the library, yet, to be unbiased when it comes to reviewing it in terms of e.g., the installation and documentation requirements via JOSS)
