### 2018-03-19 #631 @mlgill
@rasbt I am potentially interested. Could you let me know what's expected and approximate timeline? I have a hectic next two weeks due to work obligations, but it will improve after that. Thanks!
