### 2018-03-19 #631 @rasbt
Thanks @mlgill , and no worries about the time line! I think there's no official time constraint like the 30 days for other academic, reviews, but something in that ballpark would be good. Also, JOSS reviews as involved as reviews for other academic journals, because the papers are basically just a summary of the software package in 250-500 words. Mainly, the reviews is to check whether the submissions adheres to open source practices. If you should accept to review, the editor will post a checklist with instructions, e.g., similar to the one here: https://github.com/openjournals/joss-reviews/issues/634

The complete instructions are available here: https://joss.theoj.org/about#reviewer_guidelines (under the Reviewer Guidelines section).
