### 2018-03-19 #631 @mlgill
@arokem @rasbt I think this is doable -- I might get some time next week to get started, but either way will work on it in April and I think I can complete in 4 weeks. Just let me know when the checklist is ready!
