### 2018-03-19 c2e1e2a
Adds fit_params to SequentialFeatureSelector fit method
