+++ mlxtend/feature_selection/sequential_feature_selector.py
@@ -22,16 +22,17 @@ from sklearn.model_selection import cross_val_score
 from sklearn.externals.joblib import Parallel, delayed
 
 
-def _calc_score(selector, X, y, indices):
+def _calc_score(selector, X, y, indices, **fit_params):
     if selector.cv:
         scores = cross_val_score(selector.est_,
                                  X[:, indices], y,
                                  cv=selector.cv,
                                  scoring=selector.scorer,
                                  n_jobs=1,
-                                 pre_dispatch=selector.pre_dispatch)
+                                 pre_dispatch=selector.pre_dispatch,
+                                 fit_params=fit_params)
     else:
-        selector.est_.fit(X[:, indices], y)
+        selector.est_.fit(X[:, indices], y, **fit_params)
         scores = np.array([selector.scorer(selector.est_, X[:, indices], y)])
     return indices, scores
 
@@ -169,7 +170,7 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
         # don't mess with this unless testing
         self._TESTING_INTERRUPT_MODE = False
 
-    def fit(self, X, y):
+    def fit(self, X, y, **fit_params):
         """Perform feature selection and learn model from training data.
 
         Parameters
@@ -179,6 +180,8 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
             n_features is the number of features.
         y : array-like, shape = [n_samples]
             Target values.
+        fit_params : dict of string -> object, optional
+            Parameters to pass to to the fit method of classifier.
 
         Returns
         -------
@@ -248,7 +251,7 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
                 k_to_select = min_k
             k_idx = tuple(range(X.shape[1]))
             k = len(k_idx)
-            k_idx, k_score = _calc_score(self, X, y, k_idx)
+            k_idx, k_score = _calc_score(self, X, y, k_idx, **fit_params)
             self.subsets_[k] = {
                 'feature_idx': k_idx,
                 'cv_scores': k_score,
@@ -266,14 +269,16 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
                         orig_set=orig_set,
                         subset=prev_subset,
                         X=X,
-                        y=y
+                        y=y,
+                        **fit_params
                     )
                 else:
 
                     k_idx, k_score, cv_scores = self._exclusion(
                         feature_set=prev_subset,
                         X=X,
-                        y=y
+                        y=y,
+                        **fit_params
                     )
 
                 if self.floating:
@@ -298,7 +303,8 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
                                 feature_set=k_idx,
                                 fixed_feature=new_feature,
                                 X=X,
-                                y=y
+                                y=y,
+                                **fit_params
                             )
 
                         else:
@@ -306,7 +312,8 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
                                 orig_set=orig_set - {new_feature},
                                 subset=set(k_idx),
                                 X=X,
-                                y=y
+                                y=y,
+                                **fit_params
                             )
 
                         if k_score_c is not None and k_score_c > k_score:
@@ -395,7 +402,7 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
         self.fitted = True
         return self
 
-    def _inclusion(self, orig_set, subset, X, y, ignore_feature=None):
[…]
