### 2018-03-19 7365139
Adds fit_params test for SequentialFeatureSelector
