### 2018-03-21 5f7028e
Refactors travis installs to use setup.py
