### 2018-03-21 87a6cac
Replaces pip install with conda install for specific package versions
