### 2018-03-21 b8ee7d6
Removes conda install matplotlib (already included in requirements.txt)
