### 2018-03-21 a6c322a
Adds mlxtend version print
