+++ ci/.travis_install.sh
@@ -36,3 +36,4 @@ python -c "import pandas; print('pandas %s' % pandas.__version__)"
 python -c "import numpy; print('numpy %s' % numpy.__version__)"
 python -c "import scipy; print('scipy %s' % scipy.__version__)"
 python -c "import sklearn; print('sklearn %s' % sklearn.__version__)"
+python -c "import mlxtend; print('mlxtend %s' % mlxtend.__version__)"
