### 2018-03-24 8fed8ee
raise notfittederror in stacking estimators
