### 2018-03-26 #638 @rasbt
(regarding the changes in the PDF: I just added a list of acknowledgements attributing the contributors)
