### 2018-03-26 4d4dff1
Adds fit_params to ExhaustiveFeatureSelector fit method
