### 2018-03-26 2f46ec1
Adds fit_params test for ExhaustiveFeatureSelector
