### 2018-03-26 137769d
Adds missing parameter definition in fit_transform
