+++ mlxtend/feature_selection/sequential_feature_selector.py
@@ -481,6 +481,8 @@ class SequentialFeatureSelector(BaseEstimator, MetaEstimatorMixin):
         X : {array-like, sparse matrix}, shape = [n_samples, n_features]
             Training vectors, where n_samples is the number of samples and
             n_features is the number of features.
+        y : array-like, shape = [n_samples]
+            Target values.
         fit_params : dict of string -> object, optional
             Parameters to pass to to the fit method of classifier.
 
