### 2018-03-26 60da532
adding an entry to the Changelog and updating the API docs
