### 2018-04-03 fdbb375
add a new feature_importance_permutation function
