### 2018-04-08 rasbt/mlxtend#360 @mlgill
[JOSS Review] References in paper should have DOIs?
@rasbt The paper criteria mention DOIs are required for archival references, which I assume includes the references in the paper. I realize conference proceedings may not have them, but the journal articles should.
