### 2018-04-08 #638 @mlgill
@rasbt @arokem Am now able to work on this -- thanks for your patience. Have completed general checks and paper checks. Will devote some time to the functionality and documentation sections this week.
