### 2018-04-08 #638 @mlgill
@rasbt Requested DOIs links for journal article references in paper. Issue is [here](https://github.com/rasbt/mlxtend/issues/360).
