### 2018-04-09 #638 @rasbt
> Requested DOI links for journal article references in paper. Issue is here.

Thx. I added the DOIs to all papers for which they exist. The updated PDF paper is available via whedon's link above
