### 2018-04-10 c0ff0dc
fix docstrings
