+++ mlxtend/frequent_patterns/apriori.py
@@ -51,8 +51,9 @@ def apriori(df, min_support=0.5, use_colnames=False, max_len=None):
     Parameters
     -----------
     df : pandas DataFrame
-      pandas DataFrame in one-hot encoded format. For example
-      ```
+      pandas DataFrame the encoded format. For example,
+
+    ```
              Apple  Bananas  Beer  Chicken  Milk  Rice
         0      1        0     1        1     0     1
         1      1        0     1        0     0     1
@@ -62,7 +63,8 @@ def apriori(df, min_support=0.5, use_colnames=False, max_len=None):
         5      0        0     1        0     1     1
         6      0        0     1        0     1     0
         7      1        1     0        0     0     0
-        ```
+    ```
+
     min_support : float (default: 0.5)
       A float between 0 and 1 for minumum support of the itemsets returned.
       The support is computed as the fraction
