### 2018-04-11 rasbt/mlxtend#364 @mlgill
[JOSS] Link to User Guide from Examples section?
This is a suggestion, not a requirement. Mlxtend has a really nice User Guide, but it took me some time to  find it since I saw the shorter examples section first. Is it worth linking to the user guide from the examples section for those who want more information?
