### 2018-04-11 #638 @mlgill
Minor (non-critical) suggestion to link to the very nice user guide from the examples section, see [issue](https://github.com/rasbt/mlxtend/issues/364).
