### 2018-04-11 #638 @mlgill
@rasbt  I am completed with the review, other than the request to clean up documentation (which appears to be in-progress) based on the issue. When this is complete, please let me know and I will check it off.

I have also listed one other minor (non-critical suggestion) above. This isn't required.

Please let me know what else is needed other than checking off the last box. Thank you for the opportunity to review -- this was my first JOSS review and I really enjoyed it.

cc @arokem
