+++ docs/ipynb2markdown.py
@@ -54,11 +54,7 @@ def ipynb_to_md(ipynb_path):
             if line.lstrip().startswith('#'):
                 break
         for line in f:
-
-            if line.startswith('        ```'):
-                new_s.append(line[8:])
-            else:
-                new_s.append(line[4:])
+            new_s.append(line[4:])
 
     with open(md_name, 'w') as f:
         f.write(''.join(new_s))
