### 2018-04-11 rasbt/mlxtend#364 @rasbt
That's a good idea!

 I think it would make most sense to include such a link at the end of the examples section in the docstrings. E.g.,

```
    ...
    Returns
    -------
    scores : array of float, shape=(len(list(n_splits)),)
        Array of scores of the estimator for each bootstrap
        replicate.

    Examples
    --------
    >>> from sklearn import datasets, linear_model
    >>> from mlxtend.evaluate import bootstrap_point632_score
    ...
    >>> print('95%% Confidence interval: [%.2f, %.2f]' % (lower, upper))
    95% Confidence interval: [0.90, 0.98]

    More examples can be found in the user guide at
    http://rasbt.github.io/mlxtend/user_guide/evaluate/bootstrap_point632_score/
```

In addition to having helpful links in the API docs, it will also be useful if people use the `help` function in Python (or `?` in IPython).

However, the `ipynb2markdown.py` helper script should then remove this text & link from the output so that it doesn't appear on the linked website itself saying "***more*** examples can be found at ...", which could lead to some confusion.
