### 2018-04-11 #638 @rasbt
@mlgill thanks for the time doing such a thorough review, and the issues you pointed out were very helpful -- these were things I never noticed myself and I am glad that they are fixed now! Also, I am glad to hear that you liked the process, which I hope wasn't too tedious :).
