### 2018-04-15 #638 @arokem
Thanks @mlgill! Nothing more on your end.

@rasbt : I believe the paper is ready to be accepted!

It would be nice to include something to address https://github.com/rasbt/mlxtend/issues/364, but that's not a requirement. Either way, let me know when you have an archive  (e.g., a Zenodo DOI) of the version that includes all the changes following the review.
