### 2018-04-17 #638 @rasbt
Sounds great! I will try to address the suggestion in the upcoming days (prob. on the weekend) and make a new release version (and also upload it to Zenodo then), which will also include the other documentation fixes from this review. I will let you know once that's ready!
