### 2018-04-18 60ecf2c
rename refit to use_clones in stacking classifiers
