### 2018-04-19 cdfd316
fix meta feature reshuffling bug in StackingCVClassifier
