### 2018-04-21 e538cf2
add examples links
