+++ mlxtend/data/local_mnist.py
@@ -27,6 +27,9 @@ def loadlocal_mnist(images_path, labels_path):
     labels : [n_samples] numpy array
         Target class labels
 
+    For usage examples, please see
+    http://rasbt.github.io/mlxtend/user_guide/data/loadlocal_mnist/
+
     """
     with open(labels_path, 'rb') as lbpath:
         magic, n = struct.unpack('>II',
