+++ mlxtend/frequent_patterns/association_rules.py
@@ -48,6 +48,9 @@ def association_rules(df, metric="confidence", min_threshold=0.8):
       of all rules for which
       metric(rule) >= min_threshold.
 
+    For usage examples, please see
+    http://rasbt.github.io/mlxtend/user_guide/frequent_patterns/association_rules/
+
     """
 
     # check for mandatory columns
