### 2018-04-21 f1b433b
add usage example links to docs
