+++ mlxtend/classifier/stacking_classification.py
@@ -78,6 +78,8 @@ class StackingClassifier(BaseEstimator, ClassifierMixin, TransformerMixin):
         number of samples
         in training data and n_classifiers is the number of classfiers.
 
+    Examples
+    -----------
     For usage examples, please see
     http://rasbt.github.io/mlxtend/user_guide/classifier/StackingClassifier/
 
