+++ mlxtend/file_io/find_filegroups.py
@@ -48,6 +48,8 @@ def find_filegroups(paths, substring='', extensions=None, validity_check=True,
         found in the first directory listed
         in `paths` (without file extension).
 
+    Examples
+    -----------
     For usage examples, please see
     http://rasbt.github.io/mlxtend/user_guide/file_io/find_filegroups/
 
