+++ mlxtend/__init__.py
@@ -4,4 +4,4 @@
 #
 # License: BSD 3 clause
 
-__version__ = '0.12.0dev'
+__version__ = '0.12.0'
