+++ docs/sources/CHANGELOG.md
@@ -7,6 +7,30 @@ The CHANGELOG for the current development version is available at
 
 ---
 
+### Version 0.13.0dev0 (TBD)
+
+##### Downloads
+
+- [Source code (zip)](https://github.com/rasbt/mlxtend/archive/v0.13.0.zip)
+- [Source code (tar.gz)](https://github.com/rasbt/mlxtend/archive/v0.13.0.tar.gz)
+
+##### New Features
+
+-  -
+
+
+##### Changes
+
+
+- -
+
+##### Bug Fixes
+
+- -
+
+
+
+
 
 ### Version 0.12.0 (2018-21-04)
 
