### 2018-04-21 rasbt/mlxtend#364 @rasbt
Alright, I just added links to the user guides to all API docs. Thanks again for this useful suggestion!
