### 2018-04-21 #638 @rasbt
@arokem I just made the changes as suggested in [rasbt/mlxtend#364](https://github.com/rasbt/mlxtend/issues/364) and also made a new release version (0.12.0). The Zenodo link is https://zenodo.org/record/1226560

Please let me know if there's something else that needs to be done!  I am looking forward to the published version :)
