### 2018-04-21 #638 @arokem
👏 bravo @rasbt : I believe that this is ready to be accepted.

@arfon : could you please take a look and put the ribbon on this one?
